refactor(modularization): Phase 4 Step 4-S3b — governance schemas extract (ApiKey* + AuditLog* → aperag/domains/governance/schemas.py + dual-hook) (#25)
Second sub-step of 4-S3. Same pattern as 4-S3a.

6 schemas carved to governance domain: ApiKey / ApiKeyList /
ApiKeyCreate / ApiKeyUpdate / AuditLog / AuditLogList. view_models
try block extended with governance re-import (AuditLogList extends
PaginatedResponse from aperag.schema.common).

Full unit suite: 671 passed / 29 skipped.
